Refactoring SceneTreeDock::replace_node into Node::replace_by.
A portion of the code in Node::replace_by is vestigial and does not actually do anything, both it and the other portion is identical to what was in SceneTreeDock::replace_node. I have refactored the working code of replace_node in to replace_by. I suspect this duplication/copy code should actually be moved even further down the tree into Object (because it is responsible for properties, and the skipping scripts and metadata would make sense in the class responsible for those things.

I then elevated the concern of what script the resulting node should have to the "Change Type" code of SceneTreeDock, this seems like the more relevant location to make that decision. The decision is straight forward, if the new type has a script it uses that, if the old node has a script it uses that, then it gives up and uses nothing. This seems to give the best behavior for common use cases.

An improved fix for #37479